feat(subject-info): add SubjectInfo entity, DAO, service, and mapper XML
- Create `SubjectInfo` entity class with Lombok and Serializable implementation
- Define `SubjectInfoDao` interface for CRUD and custom queries
- Implement `SubjectInfoServiceImpl` to delegate operations to DAO
- Add `SubjectInfoDao.xml` with MyBatis resultMap, select, insert, update, delete statements

+++ share-club-subject/share-club-infra/src/main/java/com/shareIt/subject/infra/basic/mapper/SubjectCategoryDao.java
@@ -18,7 +18,7 @@ public interface SubjectCategoryDao {
      * @param id Primary key
      * @return The instance object
      */
-    SubjectCategory queryById(Integer id);
+    SubjectCategory queryById(Long id);
