+++ share-club-subject/share-club-infra/src/main/java/com/shareIt/subject/infra/basic/service/SubjectCategoryService.java
@@ -15,7 +15,7 @@ public interface SubjectCategoryService {
      * @param id Primary key
      * @return The instance object
      */
-    SubjectCategory queryById(Integer id);
+    SubjectCategory queryById(Long id);
